requirements.txt
ultralytics
roboflow
gradio
opencv-python
matplotlib
scikit-image
pandas
pillow
torch
torchvision
